+++ spec/rails_helper.rb
@@ -49,6 +49,8 @@ RSpec.configure do |config|
   # instead of true.
   config.use_transactional_fixtures = true
 
+  config.example_status_persistence_file_path = "#{::Rails.root}/tmp/examples.txt"
+
   # RSpec Rails can automatically mix in different behaviours to your tests
   # based on their file location, for example enabling you to call `get` and
   # `post` in specs under `spec/controllers`.